+++ ListableUI/Sources/Internal/ItemCell.ContentViewContainer.swift
@@ -137,6 +137,7 @@ extension ItemCell {
 
             guard let configuration = swipeConfiguration else { return }
 
+            let velocity = sender.velocity(in: self).x
             let offsetMultiplier = configuration.numberOfActions == 1 ? 0.5 : 0.7
             let performActionOffset = frame.width * CGFloat(offsetMultiplier)
             let currentSwipeOffset = -contentView.frame.origin.x
@@ -152,14 +153,18 @@ extension ItemCell {
             switch sender.state {
             case .began, .changed:
 
-                let swipeState = SwipeActionState.swiping(willPerformAction: willPerformAction)
-                set(state: swipeState)
+                if swipeState == .closed && velocity > 0 {
+                    // The cell is closed and this is a swipe to the right. Ignore the swipe.
+                    sender.setTranslation(.zero, in: self)
+                } else {
+                    let swipeState = SwipeActionState.swiping(willPerformAction: willPerformAction)
+                    set(state: swipeState)
+                }
 
             case .ended, .cancelled:
 
                 let swipeActionsWidth = configuration.swipeView.swipeActionsWidth
                 let keepOpenOffset = swipeActionsWidth / 2
-                let velocity = sender.velocity(in: self).x
 
                 var swipeState: SwipeActionState