+++ ListableUI/Sources/Internal/ItemCell.ContentViewContainer.swift
@@ -68,7 +68,8 @@ extension ItemCell {
             case .swiping:
 
                 let translation = configuration.panGestureRecognizer.translation(in: self)
-                xOriginOffset = contentView.frame.origin.x + translation.x
+                // No actions exist to the left, so limit overscrolling to the right to 20% of the width.
+                xOriginOffset = min(bounds.width / 5.0, contentView.frame.origin.x + translation.x)
 
                 configuration.panGestureRecognizer.setTranslation(.zero, in: self)
 
@@ -106,7 +107,7 @@ extension ItemCell {
                 insertSubview(swipeView, belowSubview: contentView)
                 swipeView.clipsToBounds = true
 
-                let panGestureRecognizer = HorizontalPanGestureRecognizer(target: self, action: #selector(handlePan))
+                let panGestureRecognizer = LeftPanGestureRecognizer(target: self, action: #selector(handlePan))
                 addGestureRecognizer(panGestureRecognizer)
 
                 swipeConfiguration = SwipeConfiguration(
@@ -150,7 +151,6 @@ extension ItemCell {
 
             switch sender.state {
             case .began, .changed:
-
                 let swipeState = SwipeActionState.swiping(willPerformAction: willPerformAction)
                 set(state: swipeState)