+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - [When swiping to delete](https://github.com/kyleve/Listable/pull/270), limit overscrolling to 20% of the cell width. This prevents undesirable visual state while maintaining swipe bounciness. Additionally, ignore initial swipes to the right which do not "open" the cell.
 
+- [Fixed a crash that occurred](https://github.com/kyleve/Listable/pull/271) when the list's width or height would become zero.
+
 ### Added
 
 ### Removed